Tentative solution to bad menus query
I traced core code as much as I could to see where the process was going off the rails... no joy.

I tried many variations of minor code and relationship tweaks to try and clear things up... no joy.

The first query to get the IDd menu works. A second query tries to work through block_id for some reason. This comes in on the ChildMenu association but I can't figure out why.

I finally tried this pushed solution. It links stepwise from Menus to ChildMenusMenus to ChildMenus. This worked.

It seems that something in the magic, in this edge case where the table does not match expectations bugs out.

We could try and send a bug report if you want.

This work around returns the entities in a different pattern than the normal belongsToMany Through but we could correct that if we want.

Or, we could carry menus in a different table. We had good reasons to share this table and I think their still valid. But maybe 5 minutes double checking would be good.