Correct the way Field output strategies are determined
The old way of having pre-configured strategies for all the use cases on a page then selecting them by model alias (the way crud data and action pattern settings are selected) turned out to be wrong.

It was much simpler to inclued a new property in the crud data object that named the output strategy that would be needed for that object. Then when the crud data was moved into the working slot, it's output configuration would simply be assembled as part of the 'saddle-up' process.